refactor: simplify example
