+++ src/transform.py
@@ -9,6 +9,9 @@ def area_circ(r_in):
     if r_in < 0:
         raise ValueError()
     area_out = np.pi * r_in**2
-    print("The area of a circle with radius r = {:3.2f}cm\
-        is A = {:4.2f}cm2.".format(r_in, area_out))
+    print(
+        """The area of a circle with radius r={:3.2f}cm is A={:4.2f}cm2.""".format(
+            r_in, area_out
+        )
+    )
     return area_out
